Diffusion compute operations are now done in the diffusion pass.
